backend/requirements.txt
certifi==2018.4.16
chardet==3.0.4
Django==1.11.13
django-cors-headers==2.2.0
django-filter==1.1.0
djangorestframework==3.8.2
httpie==0.9.9
idna==2.6
model-mommy==1.5.1
Pillow==5.1.0
psycopg2-binary==2.7.4
Pygments==2.2.0
pytz==2018.4
requests==2.18.4
six==1.11.0
urllib3==1.22
